Added cutflow table maker

+++ src/utilities.cpp
@@ -133,8 +133,8 @@ TString RoundNumber(double num, int decimals, double denom){
   if(denom==0) return " - ";
   double neg = 1; if(num*denom<0) neg = -1;
   num /= neg*denom; num += 0.5*pow(10.,-decimals);
-  int num_int = static_cast<int>(num);
-  int num_dec = static_cast<int>((1+num-num_int)*pow(10.,decimals));
+  long num_int = static_cast<long>(num);
+  long num_dec = static_cast<long>((1+num-num_int)*pow(10.,decimals));
   TString s_dec = ""; s_dec += num_dec; s_dec.Remove(0,1);
   TString result=""; 
   if(neg<0) result+="-";
